+++ StripePaymentSheet/StripePaymentSheet/Source/Categories/STPPaymentMethodParams+PaymentSheet.swift
@@ -27,4 +27,13 @@ extension STPPaymentMethodParams {
         }
         return auBECSDebit
     }
+
+    var nonnil_bacsDebit: STPPaymentMethodBacsDebitParams {
+        guard let bacsDebit = bacsDebit else {
+            let bacsDebit = STPPaymentMethodBacsDebitParams()
+            self.bacsDebit = bacsDebit
+            return bacsDebit
+        }
+        return bacsDebit
+    }
 }